refactor: :recycle: Pure Refactoring of Restaurant Management Tab View - Add/edit
