+++ components/admin/manage-restaurant-input.tsx
@@ -1,0 +1,6 @@
+/* eslint-disable @typescript-eslint/explicit-module-boundary-types */
+import { InputField } from 'components/form/inputfield';
+
+export const ManageRestaurantInput = (props): JSX.Element => {
+    return <InputField {...props} mt={5} mb={5} size="lg" />;
+};
